requirements.txt
Pillow
google-genai
streamlit